src/main/python/requirements.txt
tensorflow-probability
statsmodels
joblib
scikit-learn
scipy
seaborn
pandas
attrdict
jupyterlab
pybuilder